+++ src/main/java/com/codingchosun/backend/service/evaluation/EvaluationQueryService.java
@@ -1,4 +1,4 @@
-package com.codingchosun.backend.service.evaludation;
+package com.codingchosun.backend.service.evaluation;
 
 import com.codingchosun.backend.domain.Post;
 import com.codingchosun.backend.domain.PostUser;
